feat: add types to firestore
