--- app/firebase.ts
@@ -1,16 +1,0 @@
-import { initializeApp } from "firebase/app";
-import { getFirestore } from "firebase/firestore";
-
-const firebaseConfig = {
-  apiKey: "AIzaSyA4I6SntGUr8boAz8_UsIhoWspRrjMWuxY",
-  authDomain: "dkms-spotify.firebaseapp.com",
-  projectId: "dkms-spotify",
-  storageBucket: "dkms-spotify.appspot.com",
-  messagingSenderId: "183020545123",
-  appId: "1:183020545123:web:457c0adaf0c3855b510456",
-  measurementId: "G-GDN3X0HN0L",
-};
-
-// Initialize Firebase
-initializeApp(firebaseConfig);
-export default getFirestore();
